[PORT] Controller for the stat panel, fix runtimes and more

+++ code/game/objects/shapeshift_holder.dm
@@ -28,16 +28,16 @@
 	slink.source = src
 
 /obj/shapeshift_holder/Destroy()
-	if(!restoring)
-		restore()
-	stored = null
-	shape = null
-	. = ..()
+    restoring = TRUE
+    stored = null
+    shape = null
+    return ..()
 
 /obj/shapeshift_holder/Moved()
 	. = ..()
-	if(!restoring || QDELETED(src))
-		restore()
+	if(restoring || QDELETED(src))
+		return
+	restore()
 
 /obj/shapeshift_holder/handle_atom_del(atom/A)
 	if(A == stored && !restoring)
@@ -63,6 +63,9 @@
 		restore(knockout=source.knockout_on_death)
 
 /obj/shapeshift_holder/proc/restore(death=FALSE, knockout=0)
+	if(restoring || QDELETED(src))
+		return
+
 	restoring = TRUE
 	qdel(slink)
 	if (stored)
@@ -84,6 +87,7 @@
 		shape.mind?.transfer_to(stored)
 	if(death)
 		stored.death()
+	else if(source.convert_damage)
 	else if(stored && source.convert_damage)
 		stored.revive(full_heal = TRUE, admin_revive = FALSE)
 
@@ -92,8 +96,7 @@
 
 		stored.apply_damage(damapply, source.convert_damage_type, forced = TRUE)
 	qdel(shape)
-	if (!QDELETED(src))
-		qdel(src)
+	shape = null
 
 /datum/soullink/shapeshift
 	var/obj/shapeshift_holder/source
